feat: find products by identifier
This change replaces the `tei_urn` query parameter with a pair of `idType/idValue` parameters to filter products using all the available identifiers.

The ideal solution in this case would be to use a single parameter with schema `#/components/schemas/identifier` together with [RFC 6570-style serialization](https://swagger.io/specification/#style-examples):

```yaml
identifier-param:
  name: identifierParam
  description: If present, only the objects with the given identifier will be returned.
  in: query
  schema:
    $ref: "#/components/schemas/identifier"
  style: form
  explode: true
```

This is currently not a viable option, since limitations of the `openapi-generator` like OpenAPITools/openapi-generator#4808 would cause bugs in the generated clients.

Signed-off-by: Piotr P. Karwasz <piotr@github.copernik.eu>